Turned blur/escape window closing behaviour into composable
